# go/vt/callerid: speedup callerid context access